+++ src/plastimatch/volume.h
@@ -24,7 +24,7 @@ class gpuit_EXPORT Volume
 {
   public:
     int dim[3];		        // x, y, z Dims
-    int npix;			// # of voxels in volume
+    size_t npix;		// # of voxels in volume
 				// = dim[0] * dim[1] * dim[2] 
     float offset[3];
     float spacing[3];
